now everything appears to be working

+++ ops.py
@@ -66,7 +66,7 @@ def inference(inputs, keep_prob, data_dim, hidden1_units, hidden2_units):
         biases = tf.Variable(tf.zeros([hidden2_units]),
                              name='biases')
         hidden2 = tf.nn.relu(tf.matmul(hidden1, weights) + biases)
-        hidden2_dropout = tf.nn.dropout(hidden2, tf.constant(keep_prob))
+        hidden2_dropout = tf.nn.dropout(hidden2, tf.constant(keep_prob, dtype='float32'))
     # Linear
     with tf.name_scope('softmax_linear'):
         weights = tf.Variable(
